Add semantic conventions for log media attributes (#2022)
Logs are transmitted via several common media, such as files and
iostreams. This PR establishes semantic conventions for describing the
file or iostream from which a log was read. Notably, these attributes
are not resource attributes, as the medium used to tranmit a log is not
considered to be the originator of the log.